PD-12487 Add a basic example story
