+++ packages/CollapsibleChecklists/stories/CollapsibleChecklists.stories.js
@@ -3,7 +3,9 @@ import { storiesOf } from "@storybook/react";
 import { withKnobs } from "@storybook/addon-knobs";
 
 import BasicExample from "./examples/Basic";
+import AdvancedExample from "./examples/Advanced";
 
 storiesOf("CollapsibleChecklists", module)
   .addDecorator(withKnobs)
-  .add("Basic", () => <BasicExample />);
+  .add("Basic", () => <BasicExample />)
+  .add("Advanced", () => <AdvancedExample />);
